+++ provider/pkg/provider/client/awaiter_test.go
@@ -1,6 +1,6 @@
 // Copyright 2016-2024, Pulumi Corporation.
 
-package awsclient
+package client
 
 import (
 	"fmt"
